# customs/appointment_booking_api_doc.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:7171","description":"Generated server url"}],"tags":[{"name":"AppointmentBooking","description":"Appointment booking api"}],"paths":{"/api/v1/clients/add":{"post":{"tags":["AppointmentBooking"],"summary":"Add new client","description":"Add new client entity and its data in data source","operationId":"addClient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Client"}}},"required":true},"responses":{"200":{"description":"successful operation","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/health":{"get":{"tags":["AppointmentBooking"],"summary":"System health","description":"Check the system health","operationId":"health","responses":{"200":{"description":"successful operation","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/clients":{"get":{"tags":["AppointmentBooking"],"summary":"Fetch all clients","description":"Fetch all clients entities and their data from data source","operationId":"getClientList","responses":{"200":{"description":"successful operation","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/api/v1/clients/{id}":{"get":{"tags":["AppointmentBooking"],"summary":"Fetch client against provided id","description":"Fetch client entity and its data from data source","operationId":"getClient","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"successful operation","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["AppointmentBooking"],"summary":"Delete client against provided id","description":"Delete client entity and its data from data source","operationId":"deleteClient","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"successful operation","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"Client":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"clientName":{"type":"string"},"contactMobile":{"type":"string"},"contactMail":{"type":"string"}}}}}}